+++ core/src/main/java/io/appulse/epmd/java/core/mapper/deserializer/Deserializer.java
@@ -16,6 +16,7 @@
 
 package io.appulse.epmd.java.core.mapper.deserializer;
 
+import io.appulse.epmd.java.core.mapper.deserializer.exception.DeserializationException;
 import io.appulse.utils.Bytes;
 
 /**
@@ -25,7 +26,7 @@ import io.appulse.utils.Bytes;
  */
 interface Deserializer {
 
-  <T> T deserialize (Bytes buffer, Class<T> type) throws Exception;
+  <T> T deserialize (Bytes buffer, Class<T> type) throws DeserializationException;
 
   boolean isApplicable (Class<?> type);
 }